video_to_pdf_render/requirements.txt
Flask==3.1.1
flask-cors
yt-dlp==2025.8.22
opencv-python-headless==4.12.0.88
scikit-image==0.25.2
fpdf==1.7.2
Pillow==11.3.0
gunicorn